+++ float/src/math/mod.rs
@@ -7,6 +7,7 @@ use crate::{
     round::{Round, Rounded},
 };
 
+pub mod cache;
 pub mod consts;
 pub mod trig;